Refactor enabling services to make it reusable for all types of deployments - ssr, container, etc.
